monolayout/requirements.txt
torch==1.4.0
torchvision==0.5.0
efficientnet_pytorch==0.6.1
nuscenes_devkit==1.1.0
pyquaternion==0.9.5
numpy==1.17.0
matplotlib==3.1.1
tqdm==4.41.1
opencv_python==4.4.0.44
pykitti==0.3.1
Pillow==8.0.1
PyYAML==5.3.1
tensorboardX==2.1
